+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next release
 
+- feat: add sierra to casm class hash mapping to genesis assets
+- chore: remove ArgentMulticall from genesis assets
 - feat: remove `seq_addr_updated` from `GenesisData`
 - chore: added prometheus metrics for da layer
 - chore: bump celestia rpc crate version